Fixed couple of glitches in removing Geographies

+++ src/main/java/akressiopertti/service/GeographyService.java
@@ -115,6 +115,10 @@ public class GeographyService {
         if(geography == null) {
             throw new IllegalArgumentException("Cannot remove object with id " + id.toString());
         }
+        if(geography.getChildren().size() > 0) {
+            throw new RelationViolationException("Cannot remove Geography because it has children");
+        }
+        
         geographyRepository.delete(id);
         return geography;
     }
